fix: format code, remove test dir, and fix linting

+++ src/flag_gems/runtime/backend/_kunlunxin/ops/clamp.py
@@ -11,19 +11,19 @@ logger = logging.getLogger("flag_gems").getChild(__name__.lstrip("."))
 @pointwise_dynamic(promotion_methods=[(0, 1, 2, "DEFAULT")])
 @triton.jit
 def clamp_func_tensor(x, mini, maxi):
-    return tl.minimum(maxi, tl.maximum(mini, x))
+    return tl.minimum(maxi, tl.maximum(mini, x.to(tl.float32)))
 
 
 @pointwise_dynamic(promotion_methods=[(0, 1, "DEFAULT")])
 @triton.jit
 def clamp_func_min_tensor(x, mini):
-    return tl.maximum(mini, x)
+    return tl.maximum(mini, x.to(tl.float32))
 
 
 @pointwise_dynamic(promotion_methods=[(0, 1, "DEFAULT")])
 @triton.jit
 def clamp_func_max_tensor(x, maxi):
-    return tl.minimum(maxi, x)
+    return tl.minimum(maxi, x.to(tl.float32))
 
 
 def clamp_tensor(A, mini=None, maxi=None):
@@ -55,19 +55,19 @@ def clamp_tensor_(A, mini=None, maxi=None):
 )
 @triton.jit
 def clamp_func(x, mini, maxi):
-    return tl.minimum(maxi, tl.maximum(mini, x))
+    return tl.minimum(maxi, tl.maximum(mini, x.to(tl.float32)))
 
 
 @pointwise_dynamic(is_tensor=[True, False], promotion_methods=[(0, 1, "DEFAULT")])
 @triton.jit
 def clamp_func_min(x, mini):
-    return tl.maximum(mini, x)
+    return tl.maximum(mini, x.to(tl.float32))
 
 
 @pointwise_dynamic(is_tensor=[True, False], promotion_methods=[(0, 1, "DEFAULT")])
 @triton.jit
 def clamp_func_max(x, maxi):
-    return tl.minimum(maxi, x)
+    return tl.minimum(maxi, x.to(tl.float32))
 
 
 def clamp_min(A, mini):
